pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "busybee"
version = "0.2.0"
description = "Doctrine-driven ML starter system with explicit contracts, identity, and feedback loops"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "scikit-learn>=1.4",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "joblib>=1.3"
]
authors = [{ name = "OpenAI" }]

[project.scripts]
busybee = "busybee.interfaces.cli:main"

[tool.setuptools.packages.find]
include = ["busybee*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
